backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
apscheduler
requests
python-multipart
pyjwt
certifi
psycopg2-binary
